dom/tests/mochitest/pointerlock/test_pointerLockLost.html

<!DOCTYPE HTML>
<html>
  <!-- https://bugzilla.mozilla.org/show_bug.cgi?id=633602 -->
  <head>
    <title>Bug 633602 - file_pointerLockLost.html</title>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body onload="start();">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <div id="content">
      <div id="div">
      </div>
    </div>
    <pre id="test">
      <script type="text/javascript">
        /*
         * Test for Bug 633602
         * This test is to verify that the mouselocklost event fired by
         * the browser has the bubbles attribute set to true, and
         * the target attribute is the correct element
         */

        SpecialPowers.setBoolPref("full-screen-api.enabled", true);
        SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
                                 false);

        SimpleTest.waitForExplicitFinish(2);
        var div = document.getElementById("div")
          , mozPointerLockLostEvent = false;

        function checkStats () {
          ok(mozPointerLockLostEvent.bubbles, 
            "Manual unlock request - The mouselocklost event " +
            "should have it's bubbles attribute set to true");

          is((mozPointerLockLostEvent.target || 
              mozPointerLockLostEvent.srcElement).id, "div", "target is div");
        }
                  
        div.addEventListener("mozpointerlockchange", function () {
          document.mozExitPointerLock();  
        }, false);

        div.addEventListener("mozpointerlocklost", function(e) {
          mozPointerLockLostEvent = e;

          checkStats();
          document.mozCancelFullScreen();
          SimpleTest.finish();
        }, false);

        document.addEventListener("mozfullscreenchange", function() {
          if (document.mozFullScreen &&
              document.mozFullScreenElement === div) {
            div.mozRequestPointerLock();
          }
        }, false);


        function start() {
          SimpleTest.waitForFocus(function() {
            div.mozRequestFullScreen();
          });
        }
      </script>
    </pre>
  </body>
</html>
